requirements.txt
keyboard==0.13.5
python-rtmidi==1.5.6